Added router to main javascript app using react-router-dom
